card: split card view into section subviews
